Extract map-of-lattice logic in refinementUnit.ml (+bugfix)
Summary:
The annotation store consists of maps of refinement units, and the
lattice functionality on it has been put into `typeCheck.ml` and
`typeInferenceLocal.ml` (a bad choice) and further has been permitted
to diverge between them (which means an outright bug).

Since the underlying lattice behavior is really just how to convert
a map of a lattice into a lattice, which is already something we need
to be able to do in `RefinementUnit` code, my plan for cleaning this
up is as follows:
- extract the map-of-lattice functionality into a (private) module
  inside `RefimentUnit` **that's this commit**
- rename `refinementUnit.ml` to `refinement.ml` and move the contents
  into a module called `Unit`, so that `RefinementUnit` becomes
  `Refinement.Unit`
- create a module `Refinement.Store` and move `annotation_store`
  logic out of `Resolution` and into `Refinement.Store`. Moreover,
  move the logic that has snuck into `typeCheck.ml` and
  `typeInferenceLocal.ml` into here as well.

This will get us to the point where the annotation_store has a
single source of truth for how lattice operations work, as well
as how data access interacts with the underlying refinement units.

Reviewed By: grievejia

Differential Revision: D32530436

fbshipit-source-id: d180706372c6074978807bb2b9783a9c91a87a45