Tweak global fallback handling in resolution
Summary:
I want to move the `annotation_store` level get and set logic
into

This diff actually does appear to change some behaviors in tests, because
in most uses of `typeCheck.ml`, `global_fallback` is set to `false` and as a result
we mostly refuse to refine a local that happens to also be a global (which
happens in module and class toplevels).

The change seems correct to me. The thread-safety of refinement in a toplevel
is similar to a normal define (actually it's probably even safer - even a GIL-ectomy probably wouldn't
interleave dependent module imports), so there's not good reason why we should be
restricting refinement to not work in toplevels.

Reviewed By: grievejia

Differential Revision: D32564649

fbshipit-source-id: 7bea3754af7136844b86785b5c2ebe2fa8b1a154