Standardize less_or_equal interface
Summary:
It's been bothering me for a while that `Refinement.Unit.less_or_equal`
takes its args positionally, which is not true of almost any of
the other lattices (I'm pretty sure this is intentional: less_or_equal
takes named args to emphasize that it is asymmetric, whereas
join and meet take positional args to indicate symmetry).

Since all of the uses are now internal anyway, it's finally easy to
fix so that's what I'm doing.

Reviewed By: grievejia

Differential Revision: D32545402

fbshipit-source-id: 7f7e07fe5278da6969dbe00194a00aa8954790e7