Support site packages with is_toplevel_module field set
Summary:
Allow configurations to specify a site-package as a toplevel module, which needs to then be handled in the backend as a path with the site-root as the import root. Otherwise, the default behavior is to treat the site-package as a package with the subdirectory as the import root.

only functional difference in this diff is to pass a path in as `root$path.py` instead of only `root$subdirectory` if `is_toplevel_module` is marked.

Reviewed By: dkgi

Differential Revision: D32575380

fbshipit-source-id: a6cddc22fca43bd5cc3dcc68eb01f0840b7be60b