Avoid creating nonsense relative path from absolute path
Summary:
The existing logic doesn't seem to make sense.

We are passing in a path, called `root` to this function that is intended to gather valid paths. We call `absolute` on root, and then check if it's a directory or a file.

If it's a file, we shouldn't be calling `create_relative` on the original path and the absolut-ized version of it. What you end up with is a nonsense double path, like:

`/data/users/szhu/fbsource/fbcode/scripts/szhu/scratch/pkg/pkg_module.py/data/users/szhu/fbsource/fbcode/scripts/szhu/scratch/pkg/pkg_module.py`

I don't think we really used this function for non-directories before, but we should fix this behavior.

Reviewed By: dkgi

Differential Revision: D32603416

fbshipit-source-id: 9a5706e2b28740681283aba9635275c8708b98af