Add a create function for `CallTarget`
Summary: In a follow-up, I will add an attribute to the `CallTarget.t` record. This is currently very painful since I would need to add that attribute to all the places in tests where we create a `CallTarget.t`. Let's introduce a `create` function that creates a `CallTarget.t` with reasonable defaults, so we can easily add attributes.

Reviewed By: dkgi

Differential Revision: D32602321

fbshipit-source-id: d2d8fd7eef1f2766d64371263d10c52755f3721f