Detect implicit calls to `__call__` during call graph building
Summary:
We need to handle implicit calls to `__call__` in the taint analysis, because `__call__` is not present in the AST, and hence we do not properly infer the taint for `self`.
This diff adds an attribute to `CallTarget` to distinguish between regular calls and calls to `__call__`.
In a follow-up, we will update the forward and backward analysis to properly infer the taint on `self`.

Reviewed By: dkgi

Differential Revision: D32603665

fbshipit-source-id: 3e00871b7482b599ffc7919bac6d180d3dde9d6e