Rename the refinement integration test
Summary:
Searching by filename is harder when we have name clashes, and
now there are two `refinementTest` modules - one in analyis/test
and one in analysis/test/integration.

Renaming the integ test to "typeRefinementTest", which I think
is at least as descriptive a name and maybe more evocotive of an
integ test, fixes the problem.

Reviewed By: dkgi

Differential Revision: D32582074

fbshipit-source-id: 86dafbe283415f1dc29e6ac8ed5bd37d86e726f5