Use Refinement.Store.show for CFG dumping
Summary:
There's no reason to hand-code another manual dump of the annotation
store inside of the cfg dumping in `typeCheck.ml`, we can just use the
same `show` function that we would use for print debugging.

Once again, this cleans up some code that was breaking the abstraction
boundary by deconstructing the store; we don't want modules to be able
to reach into the guts of Refinement.Store like that. Plus, it's just less
code to maintain.

Reviewed By: shannonzhu

Differential Revision: D32583128

fbshipit-source-id: 8bc2fb0507008ece9a831b63bc72f5441ff46a0d