(bugfix, 2/2) fix temp-vs-not and new-vs-refine bugs in refinement
Summary:
This is a rather large commit, but it has to be done in one
shot to make any sense (in theory I could have done two
separate bugfixes, but the problem is that the two bugs
I am fixing interact with one another in very confusing ways
in the tests, so it's harder both to write and review the code
if I try to break this up).

The TL;DR is that this fixes two issues:
- we need to distinguish between setting new
  values (which is what we do when handling assignments and any
  AST operation that adds a name to our scope (which potentially
  masks some preexisting name)) versus refining existing values.
- Moreover, we need to be considerably more disiplined in handling
  temporary versus non-temporary attributes than we were before.

**A slightly more detailed summary of temp vs non-temp issues**

Basically, we need to make sure that temporary refinement
data forms a consistent "mask" over non-temporary refinement
data.
- any operation on temporary data doesn't impact non-temporary
  data at all
- any operation causing a new temporary refinement should copy
  any existing non-temporary refinement
- any operation on non-temporary refinements needs to be
  applied to *both* the non-temporary refinement tree and,
  *if it exists already*, any temporary refinement tree
  - because otherwise, the mask becomes inconsistent and
    temporary refinements will "hide" subsequent non-temporary
    refinements. This is what caused the weird bug where the
    order of isinstance checks led to different reveal_type
    behaviors.

**A slightly more detailed summary of new vs refine operations**

The new versus refinement difference means we need to have 2
different functions in resolution:
- new_local_with_attributes: destroys any attribute subtree
  currently associated with this name.
- refine_local_with_attributes: updates the Annotation.t
  associated with some name, but leaves the subtree intact

We don't actually need to have `refine_local` or
`new_local` for correctness, because the with attributes versions
are capable of handling a name that exactly matches an existing
local. But, the `partition_name` function is expensive so we
want these functions to exist for the sake of optimization. As
a result we wind up with four variations.
- new_local and refine_local, which skip the partition_name call
- new_local_with_attributes and refine_local_with_attributes,
  which work on both base locals and attribute accesses but
  require a `partition_name` call.

**Changes in `refinement.ml` to support new vs refine**

At the `Refinement.Unit` level, I needed to change the name
of the `set_attribute` and `get_attribute` functions to
`set_annotation` and `get_annotation`, and add a doc comment
clearly explaining that these operations can affect either the
root or some attribute (this is actually undoing a previous
change I made, because in the absence of a docs comment it wasn't
clear to me that this operation can possibly set the base).

At the `Refinement.Store` level this requires us having a
`wipe_subtree` flag on some operations: when creating a new
refinement this should be set to true and when refining an
existing refinement this should be set to false.

**Changes to `typeCheck.ml` and `typeInferenceLocal.ml`**

Basically operations that used `set_local` and `set_local_with_attributes`
have to be triaged into either `refine_.*` or `new_.*` operations.

In `typeCheck.ml` we want to use the `new_` operations for
anything creating a new binding: assignments, parameters and
lambda parameters, names inside of generator expressions, etc.
We want to use `refine` mainly for asserts.

In `typeInferenceLocal.ml` we want to use `refine_` pretty much
everywhere (this code is proxying to `typeCheck.ml` for all the
cases that involve new bindings).

Reviewed By: shannonzhu

Differential Revision: D32581454

fbshipit-source-id: ac2eb068cc0a83c5f0f3f01f5fc55a605dc7fcc1