Clean up a block where typeCheck mutated the store
Summary:
I'm trying to preserve the abstraction layer that `Refinement.Store` is
an implementation detail in `Resolution`, and all static analysis code
relies on the `Resolution` layer - this makes it *vastly* easier to
fully understand refinement logic and make changes as needed.

One of the really hard to follow blocks was where param information was
added into the resolution in `typeCheck.ml` - here we created an
almost 1000-line let binding defining `(errors, annotation_store)`,
and then had a one-line step where we reconstructed the `resolution`
out of that `annotation_store`.

This was (a) violating the abstraction boundary in a big way, all
of the operations ought to have been going through the
`Resolution.new_local` function which is intended for handling new
bindings, and (b) unnecessarily convoluted: I verified that the
`resolution` was never re-bound, so passing the raw store around
is not necessary.

The only tricky thing is that to avoid accidentally allowing default
values to make use of previous parameters we need to keep the original resolution
distinct, inside the folding of check_parameter, from the new resolution
we are building up (as well as from any modifications for args and kwargs).
I do this by always binding the resolution resulting from adding a parameter
to `new_resolution`; since the code always resolves things using `resolution`
there's no risk of contamination; keeping things correct does *not* require
breaking the abstraction boundary!

Reviewed By: shannonzhu

Differential Revision: D32583074

fbshipit-source-id: a6cfc5414348aa746e74e36ab484cbff73110bb8