Move `meet` logic down to where it belongs
Summary:
It's pretty crazy for `typeCheck.ml` to define a `meet`
operation for `Refinement.Store` inline.

This commit moves `meet` to `Refinement.Store`, and adds
a `Resolution.meet_refinements` wrapper function so that we don't have
to jump across layers of the code.

The context where we do a proper `meet` is Boolean `and` operations,
which leads to us taking the union of all restrictions (therefore
the greatest lower bound of the type lattices).

Reviewed By: dkgi

Differential Revision: D32582460

fbshipit-source-id: 88a38807bd95b210dc702c70d23010a7bbe108a1