Add `update_existing_refinements`, hide innards
Summary:
The `typeInferenceLocal` code had some low-level mangling
of the guts of `Refinement.annotation_store` that it was
using to perform an update-existing-refinements operation.

Here we provide that functionality in `Refinement` and
`Resolution` where it belongs, and call into that code
from `typeInferenceLocal.ml`.

In this commit, I'm able to remove the functions
`Resolution.annotations` and `Resolution.temporary_annotations`,
which as far as I'm concerned were entirely in violation of the
abstraction boundaries we want to be enforcing between
`Refinement | Resolution | TypeCheck`.

Reviewed By: shannonzhu

Differential Revision: D32583976

fbshipit-source-id: c266d431cd60513bcd4fc65e86f032ff292fcd9f