Add get / set to Refinement.Store
Summary:
This commit migrates the `Refinement.Store`-level handling
of getting and setting annotations from `Resolution` to
`Refinement`. It should leave logic exactly as before.

NOTE: there's a bug I noticed here, which I've faithfully
reproduced since this diff is just a refactor, but I'll squash it in
the next commit: in the `set_as_base` function (which is reproducing
the relevant logic from `Resolution.set_local`), if the `temporary`
flag is not set we'll just update `annotations`. But that means
the refinement actually gets masked by any temporary refinements! Again,
I do want to fix that bug but it will require some care so I'll do it
in a later commit.

Reviewed By: shannonzhu

Differential Revision: D32568714

fbshipit-source-id: e7f1c9dba704446c3fcc38ee0f474c73eb5f441e