add `has_nontemporary_annotation`
Summary:
Deep in typeCheck.ml we were defining an `is_local` flag that was
reaching into the guts of the annotation_store; this is the last
big place where `typeCheck.ml` was reaching into the guts of the
annotation maps to do things manually.

I don't think `is_local` is actually a safe name for this, because
usually "local" means a top-level local in the Resolution and
Refinement code; I think this function could apply to an attribute
as well. So I named it `has_nontemporary_annotation`.

Reviewed By: shannonzhu

Differential Revision: D32593137

fbshipit-source-id: 13ba40e7864d0bf1206bbdc082875c786f41f6b0