Fix TypeInferenceLocal.State.pp
Summary:
I was dealing with a bug in a previous commit where I
accidentally set a temporary annotation in addition to
a non-temporary annotation where I hadn't intended to,
and I got this error message:
```
expected:   Expected return: typing.Any
  Types:
    x -> [Base: (int: m); Attributes: []]
  Errors:

but got:   Expected return: typing.Any
  Types:
    x -> [Base: (int: m); Attributes: []]
  Errors:

differences:
```

Confusing? Yes.

The problem is that we had a hand-written pp that ignored
temporary annotations entirely, so I couldn't see the
difference that was failing the test.

We shouldn't hand-roll a half-baked attempt to print the
resolution, we should just print it.

Reviewed By: dkgi

Differential Revision: D32581941

fbshipit-source-id: bdf2fae5f09a61bc5c6338171fb81e8353497fd7