(bugfix, 1/2) Add a test showing weird refinement behavior
Summary:
This diff is a pre-bugfix diff, recording how temporary
annotations masking non-temporary annotations currently
produces a bug where refinement of a variable can fail
whenever an attribute has previously been refined.

What's going on in this test is that in the second if,
we first refine `a.x` to `int` in `temporary_annotations`
and then we set `a` to `B` in `annotations`, then we'll
never see the refined view of `a`.

Reviewed By: dkgi

Differential Revision: D32574202

fbshipit-source-id: 0367e1c699f2219607d51cf4f3662cab7c8bf1ae