Type restrictions for value patterns
Summary:
# This diff

The value and singleton patterns implicitly also impose typing restrictions.

Singleton pattern was already working, because it was being translated using `is`, and we handle type refinement with `is`.

The `value` in a value pattern can only be a literal/constant*, so using `is` is equivalent. This diff changes translation of `value` to use `is`.

*The reason the AST uses an expression is that it could either be a one of builtin types' constant (which maps to `Expression.Constant` in the AST) or an enumeration constant (which maps to an attribute in the AST). We add a test in here to make sure refinement is working for enums as well.

# Broader context

See D32027190 (https://github.com/facebook/pyre-check/commit/0d16c9b5206fefc2d00327df9078703eef93f76f).

Reviewed By: pradeep90

Differential Revision: D32624003

fbshipit-source-id: a72c247099c9abbba1d9d428f5e96279e0e250d6