Make MalformedConfiguration errors typed
A malformed configuration exception occurs when either the taint.config
file is not found or it is not a valid json file. During the validation
of taint.config file with the json specs, we raised an exception if it
was illegal.

Now added code to pretty print the exception via Log.error and specify
the line number and column number along with the file name.

Since YoJSON doesn't expose or store line numbers, this is implemented
by creating a custom Bi_output buffer and specifying it to YoJSON so
that YoJSON uses it for parsing and the last line of the contents would
be the line in which the error occured and the position and length of
such the last line would give us the line number and column number of
the error.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>