Support `__getattr__` in protocol refinement
Summary:
To unblock pytorch (and probably some other OSS projects that use
a lot of dynamic python), we need to be able to handle classes
that implement attributes via `__getattr__`.

We currently ignore `__getattr__` when resoloving protocols, but this
doesn't work well for dynamic code - it's entirely possible for
a class that has a `__getattr__` returning `Any` to implement any
protocol.

There's obviously no type safety, but we already happily accept
any operation whatsoever (*other* than passing to a function that
expects a protocol) on such types. We should be consistent and
flexible rather than erring on the side of safety when working with
dynamic python which cannot be done in a type-safe way.

Reviewed By: shannonzhu

Differential Revision: D32697011

fbshipit-source-id: 82ea322b9e76e1b19b7c1e57367bf85a798f7319