Add an edge from the split block to the `finally` block in the control flow graph
Summary:
# Problem
We currently completely ignore the `finally` block if the `try` block always returns or raises an exception.

For instance:
```
def baz() -> None:
    try:
        return foo()
    finally:
        bar() # ignored by pyre and pysa.
```

# Solution

The problem is in the code that builds the control flow graph for the try block. In that specific case, the `finally` block has no predecessor at all.
As a workaround, let's add an edge from the split block (i.e, before the `try`) to the `finally` block, so we always analyze it.

Note that this is still not perfect. In terms of taint flow analysis, we are still missing cases where a variable is tainted within `try` and used in `finally`.
In theory, we would need to add edges from all statements that can throw exceptions into the `finally` block.

Reviewed By: kbansal

Differential Revision: D32634248

fbshipit-source-id: cb67aa03277c729c95eabda33a16c4fb3cdcc88d