Rename variable.
Summary:
In assertions with forms like `assert type(x) != t`,
we are using `expected` as the name of `t`.

I sort of get where this is coming from, often in assertions
we're unit testing - but we usually wouldn't say the thing we're
saying we *shouldn't* have is "expected"!

Anyway, I decided to call this `mismatched_type` which
seems more descriptive

Reviewed By: grievejia

Differential Revision: D32722132

fbshipit-source-id: 8be80b833e7a6d3a091348edd963b7a5bf41d671