Add test illustrating unsound refinement
Summary:
We aren't properly verifying that the generic type is iterable
before we refine based on `all`, which could matter in edge cases
where a type is generic and is also iterable, but the iteration isn't
over it's generic type parameter.

I'll stack a commit on top fixing the bug, it's nice to put the
test up first to confirm that it in fact verifies the bug / confirms
a fix.

Reviewed By: grievejia

Differential Revision: D32738557

fbshipit-source-id: 3057721e4916ba5905e1e1547625287cb3600b94