Track property calls that can also be regular attribute accesses in the forward analysis
Summary:
# Problem
There are cases where an attribute access should be considered both a property call and a regular attribute access.
For instance:
```
class A:
  property
  def foo(self): ...

class B:
  pass

def test(x: Union[A, B]):
  return x.foo
```
We currently consider this a call to A.foo and ignore that B.foo is a regular attribute.

# Solution
Now that we properly store the information in the call graph, let's handle it in the forward analysis.

Reviewed By: dkgi

Differential Revision: D32712053

fbshipit-source-id: a8b82b7b6d2a4d5c139f74ce28883194705d2314