Detect wether a call to a property should also be considered a regular attribute access
Summary:
# Problem
There are cases where an attribute access should be considered both a property call and a regular attribute access.
For instance:
```
class A:
  property
  def foo(self): ...

class B:
  pass

def test(x: Union[A, B]):
  return x.foo
```
We currently consider this a call to `A.foo` and ignore that `B.foo` is a regular attribute.

# Solution
We want the call graph to tell us whether an attribute access can also be a regular access or not.
In a follow-up, we will update the forward and backward analysis to handle both cases.

Reviewed By: dkgi

Differential Revision: D32709435

fbshipit-source-id: dddee27f8782ebbc655eb4e6550c57f2a4648c2c