Store object targets for attribute accesses in the call graph
Summary:
One performance bottleneck of the analysis is actually to resolve types.
Type resolution is queried everytime we see an attribute access, in order to fetch possible object models, i.e models for global objects (e.g, `os.environ`) or attribute models.

As a performance optimization, let's store object models in the call graph. This way, we only need to resolve types during call graph building (it already does it anyway). In a way, it makes sense to store it in the call graph anyway, to show the explicit dependency between models.

In a follow-up, we will fix the taint analysis to use the call graph information.

Reviewed By: dkgi

Differential Revision: D32747764

fbshipit-source-id: 3ffb08a423c26af03f5ce14b853a1cf0f2496195