Track property calls that can also be regular attribute accesses in the backward analysis
Summary:
# Problem
There are cases where an attribute access should be considered both a property call and a regular attribute access.
For instance:
```
class A:
  property
  def foo(self): ...

class B:
  pass

def test(x: Union[A, B]):
  return x.foo
```
We currently consider this a call to A.foo and ignore that B.foo is a regular attribute.

# Solution
Now that we properly store the information in the call graph, let's handle it in the backward analysis.

Reviewed By: dkgi

Differential Revision: D32729555

fbshipit-source-id: 194ab555bfbe4a60b15e9b78689f7c7dbf75aa4c