Extract function to check arguments against a given signature.
Summary:
This function essentially does two things: (a) get the parameter-argument mapping (b) check each argument against the parameter type.

Ideally, we should be able to replace this with (a) piped into (b).

But a function that accepts a ParamSpec callable and also accepts `*args, **kwargs` needs to call this function recursively. This is because it needs to solve constraints for the callable *before* it can get the precise parameter-argument mapping for `*args, **kwargs`.

So, extracting this function as is for now.

Renamed the function.

Reviewed By: grievejia

Differential Revision: D32840070

fbshipit-source-id: e5e4d8c9bee81b8ef49d551ca0d4a933d5f69a60