Split `model.ml` into multiple files
Summary:
The source `model.ml` was a dumping ground for everything related to models. The worst is that it was not actually defining a model.
Let's split this into multiple files:
* `callModel.ml[i]` which defines a call model, i.e a model instantiated at a call site.
* `classModels.ml[i]` which infers class models before the analysis, for instance for NamedTuple.
* `modelParser.ml` gains the `get_source_models` function to fetch all taint stubs.

This also cleans-up `Taint.Model` which was an aggregate of different modules defined in `taint.ml`. This was really confusing as you would expect `Taint.Model` to be defined in `model.ml`.

Reviewed By: dkgi

Differential Revision: D32858011

fbshipit-source-id: c079a8ee1a9edf4506f95e44a5bdfda8579536f5