Rename `sig_t` to `instantiated_return_annotation`.
Summary:
Problem: The type represents the result of signature selection. We either find a return annotation and instantiate it (if it is generic) or we fail to find it, in which case we return the closest matching annotation along with the reasons for failure.

I'm just going with `instantiated_return_annotation`, since that's better than `sig_t`.

Bikeshedding welcome.

abbr

Reviewed By: dkgi

Differential Revision: D32840158

fbshipit-source-id: 4c89a55b10d745de4565e8365e2d797250cee70e