Refactor API for `join_return_annotations`
Summary:
Problem: We were doing a lot in `resolve_signatures`.

It's a code smell to accept a sole `Optional` parameter. Cleaner to accept the type you care about and pipe in the `Optional` monad using `>>=`.

The parameter names aren't great. That's because we're sending in a bunch of info to the function. Will clean up later if possible.

Reviewed By: dkgi

Differential Revision: D32899381

fbshipit-source-id: dbe2490f58161d19fb95861ac295209ac01d42d1