Add always breadcrumbs on tito specified by models on obscure callables
Summary:
Using `TaintInTaintOut[Via[feature]]` on an obscure function (i.e, without code) leads to a may feature instead of an always feature. This is because the analysis considers two propagations: one from the user-specified model (with the feature) and one from the obscureness (without the feature).

This diff properly adds an always feature by turning the obscureness into an implicit propagation during the analysis of a call.

Reviewed By: dkgi

Differential Revision: D32961835

fbshipit-source-id: 159b1a4bcb2b7dfa1f32e14ea9725534affb884a