Make `get_callables` return a list of callable data.
Summary:
Problem: We want to resolve `arguments` in the presence of the respective callable so that we can use the parameter types. Right now, we have one `arguments` that we resolve independent of the callables.

This diff: Pair each callable with its own set of arguments. Right now, they're all just the same previously-resolved argument *types*.

Later:
+ Make `get_callables` return a list, not optional list.
+ Make the arguments be argument *expressions*.
+ Modify the `callable_data` after some further refactoring.
+ Resolve argument expressions using parameter types.

Reviewed By: dkgi

Differential Revision: D32973450

fbshipit-source-id: d7e256d419910fd5e39d01b0e312c3568b29d7fa