Carry arguments along with the selected return annotation.
Summary:
Problem: In one case when a callable is mismatched with arguments, we need to raise an error using `arguments`. But we are moving to a per-callable resolved `arguments` instead of a common `arguments`.

So, carry along the `arguments` for that particular callable.

Design choice: I went with a type parameter for `callable_data`. It is `unit` when there is no return annotation.

Reviewed By: grievejia

Differential Revision: D33063897

fbshipit-source-id: 4306752eeff482357ab9b8096eb3d30c186f838c