Add variant to represent unknown callable attribute.
Summary:
Problem: We want to raise "undefined attribute" error when neither `foo.__add__` nor the inverse operator is present. Currently, we do this by just looking at the `arguments` local variable. With bidirectional inference, we won't have access to `arguments`. So, we need to preserve the attributes that were not present.

This diff adds a variant to preserve that data when getting callables from the callee type. The variant is not consumed anywhere.

The next diff will replace the current logic for "undefined attribute" error using the above variant.

Reviewed By: grievejia

Differential Revision: D33063890

fbshipit-source-id: 842d39e65ee475b2663534cb297e3325bb520a12