Add basic variant for known callable types.
Summary:
Problem: Need to add a variant to handle the case where `foo.__add__` is not a callable type, in which case we need to look up the inverse operator.

Add a basic module with the default `KnownCallable` variant.

Moving the type definition down because we need to use the `callee_attribute` type next.

Reviewed By: grievejia

Differential Revision: D33063892

fbshipit-source-id: bf9d3f853522c30e2f121f8f8d9984911d950eb0