Remove direct dependency on `arguments` when computing undefined attribute error.
Summary:
Problem: When `foo.__add__` doesn't exist, we check whether the inverse operator exists. If that doesn't work out either, we need to raise either "unsupported operand" or "undefined attribute error". We were doing that by looking at the common `arguments` variable. But that will not be available any more.

So, this diff uses the `UnknownCallableAttribute` variant to compute the above errors.

Remaining: Remove the last dependency on arguments in `get_callables`.

Reviewed By: grievejia

Differential Revision: D33063900

fbshipit-source-id: df3279aa76ef9b6d531cc1f6c5f36f59bf061b26