Implement local breadcrumbs
Summary:
This diff implements local breadcrumbs. It basically stores breadcrumbs inferred in the current callable and breadcrumbs propagated from a call in different sets.
This allows us to show the breadcrumbs inferred in each separate frame, which helps triaging issues.

As for the implemetation: the local breadcrumb set is stored in the `LocalTaint` module, next to tito positions, since that set only depends on the callee + call position, and not on the source or sink kind. When applying a call, inferred sets and propagated sets are joined.

Note that for now we are not handling first indices and first fields, as they will need special logic when merging. This will be handled in a follow-up.

Reviewed By: fahndrich

Differential Revision: D33037819

fbshipit-source-id: d854cc91c9226f554246d8cda0d5c6f764a9d811