Local breadcrumbs for first fields and indices
Summary:
This diff implements local breadcrumbs for first fields and indices.
This requires special logic since joining propagated and local first fields/indices is not a simple union of sets.

Note that with this approach, we are getting field fields and indices on each frames, and they might not be propagated if the source or sink already has field fields/indices.

Reviewed By: fahndrich

Differential Revision: D33079837

fbshipit-source-id: c2d8f48f2f52c5f7a32a8d91f62fbdb1bf608cd0