Inline signature selection functions.
Summary:
Now that we have callables before getting arguments, we want to resolve argument using the callable type and then call signature select.

This diff inlines the signature selection code. Right now, we only use bidirectional signature selection if there is a single callable. We fall back to the old signature selection for a union of callables. Otherwise, we would have to deal with the perf consequences of resolving arguments multiple times; I'll do that after getting it working end to end.

Remaining:
+ Resolve arguments using callable type.
+ Read the configuration field for constraint-solving style. Not able to access `configuration` directly right now.

Reviewed By: grievejia

Differential Revision: D33141475

fbshipit-source-id: 5c8171e9dddaa7212a5f0eddad06c456480382d9