Remove named parameter for `forward_expression`.
Summary:
Problem: Want to add an optional parameter `expected_type` so that we can gradually pass in the expected type from different callers. But that is hindered by the fact that all the parameters are named parameters.

In any case, having the last parameter be positional help us curry easily. RFC.

Reviewed By: grievejia

Differential Revision: D33163706

fbshipit-source-id: c836f0db7263bc74f0a66faa28178c6773bc71ef