Decouple `arguments` from `get_callables`.
Summary:
Problem: Inverse operator computation looks at the argument types whereas we want to call `get_callables` before we get the argument types.

Factor out some other parts of `get_callables` that depend on `arguments`.

Basically, there are two phases:
1. before application - where we have to compute callables without access to the resolved `arguments`
2. application - apply the callables to `arguments` resolved for each callable.

Design choice: I made `CallableApplicationData.t` generic in `arguments`. Using `option` wouldn't work because it is guaranteed that `arguments` are present in the second phase. So, in the first phase, the type argument is `unit`; in the second phase, the type argument is `Argument.t list`.

Note that this is not leaked out into the code, so readers don't need to worry too much about it.

This was the last instance of `arguments` being captured implicitly by a local function. Phew! Now, we can proceed with bidirectional inference.

Reviewed By: grievejia

Differential Revision: D33141477

fbshipit-source-id: 2f08c40e5fed00d964e456b4179392a19e6c1d89