Revert D32815758 and D32816768
Summary:
We can simplify things quite a bit with change of switching the definition of coverage as discussed in D33100198.

One of those is removing AnnotationKind in coverage_collector (we only care about functions). The simplest way is to revert the diffs that added them -- luckily, the changes since have been simple enough that the reverse patches succeed.

Reviewed By: grievejia

Differential Revision: D33147607

fbshipit-source-id: e67c2ce20aec92a3d5699a0d5c5dec23beb5f4a4