Report 100% coverage for strict files (1/2)
Summary:
This diff:
Per the new definition, files are always covered if they are strict. Does it for the CLI tool.

Whats to follow:
Only partly done. Still need to do on the VSCode side, which requires a bit more work as the configuration is not handy at the current entry point. Will do that in a separate diff.

Context:
See D33100198.

Reviewed By: grievejia

Differential Revision: D33114161

fbshipit-source-id: df27c26dcab8af2c5c08a908555300d75a2886e7