Respect excludes in VSCode as well
Summary:
# This diff

Respect excludes in VSCode when computing coverage.

Note that this doesn't account for `targets`, which requires a completely different model (we need to switch from client side to server side computation). In the meantime, let's at least be consistent with `coverage` and `statistics` commands (which also respect `exclude`, but not `targets`).

Reviewed By: grievejia

Differential Revision: D33201271

fbshipit-source-id: ee12a2fa9ca044b4fed92e600fd7b2a80320eb2e