Better error message when Union member has an attribute error
Summary: We see user confusion when one element of a Union is causing an error, but we don't mention that we pulled that type out of the Union and are not mentioning the rest of the Union. Has led users to believe we were misinterpreting the type. See task for more details.

Reviewed By: grievejia

Differential Revision: D32782392

fbshipit-source-id: 2ee441d5e7205cca2c7aa4b43c9ef963d7c3918a