Add unit tests for getting a parameter-argument mapping.
Summary:
Problem: Right now, we hackily unpack `args` in `foo(x, *args)` if we know that it is a concrete tuple type.

However, that doesn't work with bidirectional inference, where we don't have the resolved type when mapping arguments to parameters.

Will get rid of the hack in upcoming diffs.

Reviewed By: dkgi

Differential Revision: D33248016

fbshipit-source-id: 8cc4ce7567657463bc8b2bc04497670fe2eb708b