Allow spreading `*xs` across parameters.
Summary:
Problem: We were hackily unpacking `*xs` if it had a concrete tuple type. For other tuple types such as `Tuple[str, *Tuple[int, ...]]`, we used to find the common type (i.e., `object`) and check that against the parameter type. This led to spurious errors.

This diff uses the index into the `*xs` to find out the type within it that should be assigned to the current parameter.

Still no change in external behavior because the previous stages in the signature selection pipeline don't yet pass in the index into the starred argument.

What does this have to do with function-level constraint-solving, you ask? The old hack prevented us from delaying the point at which we resolve arguments. That is the only connection.

Remaining: Still need to handle the case where the parameter type is `*args: ...`. Next diff.

Reviewed By: stroxler

Differential Revision: D33299313

fbshipit-source-id: 362c778256d8e2c293ba6ef1dbf86f5299b4b5eb