Allow checking the suffix of a `*xs` against an expected `*args`.
Summary:
See D33299300 for more details.

When checking arguments against parameters, if we are told that only part of `*xs` is mapped to a `*args` parameter, then we should slice the type of `*xs` as needed.

Note that this isn't used yet because the previous stages of signature selection don't yet do partial-mapping of `*xs` against `*args`.

Reviewed By: stroxler

Differential Revision: D33299311

fbshipit-source-id: b488425b20914a2ae740cc9fb1bd38b81f781011