Add field to track index into starred arguments.
Summary:
Problem in this and next diffs: We want to map an unpacked argument `*xs` to the parameters. Right now, we hackily support concrete tuples by using the resolved type and unpacking it manually. But we won't have the resolved type at this stage in bidirectional inference.

The main problem is that, without the type, we don't know exactly which elements of the `*xs` map to which parameters. Right now, we assume that it could be *any* element, so we essentially find the union of the element types and check if it can be assigned to the parameter. This is incorrect.

To distribute `*xs` without using the type, whenever there is a starred argument, I will wrap the remaining positional arguments into a tuple.

```
def foo(y1: int, y2: int, y3: int, y4: int, *ys: int, d: int) -> None: ...

foo(a, *xs, b, c, d=2)`
```

So, `foo(a, *xs,  b, c, d=2)` will wrap `synthetic_tuple = (*xs, b, c)`. It will then assign this tuple to parameters using indexes. That means we map:

```
y1 -> a
y2 -> synthetic_tuple[0], if it exists; else use the default value for y2 (if any)
y3 -> synthetic_tuple[1]
y4 -> synthetic_tuple[2]
ys -> synthetic_tuple[3:]
```

There are some edge cases with named arguments. We don't handle them correctly now anyway. Will fix as much as possible.

This diff just adds the field to track the index into the tuple.

RFC on a better data type design.

Reviewed By: stroxler

Differential Revision: D33299300

fbshipit-source-id: 69ae3cf4a94ced89182b22c7dca443a0e5d6b814